assistant/requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==6.1.1
py-cpuinfo==9.0.0
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sounddevice==0.5.1
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.73
ultralytics-thop==2.0.14
urllib3==2.3.0
